Add a `MutableSettings` type that contains settings that can change at runtime.
This is just the first part of a settings overhaul to more gracefully handle the fact that these settings can change at runtime. For now, there's no practical change, this is just a minimal first step to keep the changes atomic and reviewable.
