+++ tracer/src/Datadog.Trace/Configuration/MutableSettings.cs
@@ -1,0 +1,753 @@
+﻿// <copyright file="MutableSettings.cs" company="Datadog">
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
+// </copyright>
+
+#nullable enable
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using Datadog.Trace.Ci;
+using Datadog.Trace.Ci.CiEnvironment;
+using Datadog.Trace.Configuration.ConfigurationSources.Telemetry;
+using Datadog.Trace.Configuration.Telemetry;
+using Datadog.Trace.Processors;
+using Datadog.Trace.Tagging;
+using Datadog.Trace.Util;
+
+namespace Datadog.Trace.Configuration;
+
+/// <summary>
+/// Settings that can change during the lifetime of the application. This can include settings updated in
+/// code or via remote configuration. Note that the specific instance is immutable, but there may be a
+/// new version in the lifetime of the application
+/// </summary>
+public class MutableSettings
+{
+    // we cached the static instance here, because is being used in the hotpath
+    // by IsIntegrationEnabled method (called from all integrations)
+    private readonly DomainMetadata _domainMetadata = DomainMetadata.Instance;
+
+    internal MutableSettings(
+        IConfigurationSource source,
+        IConfigurationTelemetry telemetry,
+        OverrideErrorLog errorLog,
+        TracerSettings tracerSettings)
+    {
+        Telemetry = telemetry;
+        ErrorLog = errorLog;
+        var config = new ConfigurationBuilder(source, Telemetry);
+
+        LogsInjectionEnabled = config
+                              .WithKeys(ConfigurationKeys.LogsInjectionEnabled)
+                              .AsBool(defaultValue: true);
+
+        var otelTags = config
+                      .WithKeys(ConfigurationKeys.OpenTelemetry.ResourceAttributes)
+                      .AsDictionaryResult(separator: '=');
+
+        Dictionary<string, string>? globalTags;
+        if (tracerSettings.ExperimentalFeaturesEnabled.Contains("DD_TAGS"))
+        {
+            // New behavior: If ExperimentalFeaturesEnabled configures DD_TAGS, we want to change DD_TAGS parsing to do the following:
+            // 1. If a comma is in the value, split on comma as before. Otherwise, split on space
+            // 2. Key-value pairs with empty values are allowed, instead of discarded
+            // 3. Key-value pairs without values (i.e. no `:` separator) are allowed and treated as key-value pairs with empty values, instead of discarded
+            Func<string, IDictionary<string, string>> updatedTagsParser = (data) =>
+            {
+                var dictionary = new ConcurrentDictionary<string, string>();
+                if (string.IsNullOrWhiteSpace(data))
+                {
+                    // return empty collection
+                    return dictionary;
+                }
+
+                char[] separatorChars = data.Contains(',') ? [','] : [' '];
+                var entries = data.Split(separatorChars, StringSplitOptions.RemoveEmptyEntries);
+
+                foreach (var entry in entries)
+                {
+                    // we need Trim() before looking forthe separator so we can skip entries with no key
+                    // (that is, entries with a leading separator, like "<empty or whitespace>:value")
+                    var trimmedEntry = entry.Trim();
+                    if (trimmedEntry.Length == 0 || trimmedEntry[0] == ':')
+                    {
+                        continue;
+                    }
+
+                    var separatorIndex = trimmedEntry.IndexOf(':');
+                    if (separatorIndex < 0)
+                    {
+                        // entries with no separator are allowed (e.g. key1 and key3 in "key1, key2:value2, key3"),
+                        // it's a key with no value.
+                        var key = trimmedEntry;
+                        dictionary[key] = string.Empty;
+                    }
+                    else if (separatorIndex > 0)
+                    {
+                        // if a separator is present with no value, we take the value to be empty (e.g. "key1:, key2: ").
+                        // note we already did Trim() on the entire entry, so the key portion only needs TrimEnd().
+                        var key = trimmedEntry.Substring(0, separatorIndex).TrimEnd();
+                        var value = trimmedEntry.Substring(separatorIndex + 1).Trim();
+                        dictionary[key] = value;
+                    }
+                }
+
+                return dictionary;
+            };
+
+            globalTags = config
+                        .WithKeys(ConfigurationKeys.GlobalTags, "DD_TRACE_GLOBAL_TAGS")
+                        .AsDictionaryResult(parser: updatedTagsParser)
+                        .OverrideWith(
+                             RemapOtelTags(in otelTags),
+                             ErrorLog,
+                             () => new DefaultResult<IDictionary<string, string>>(new Dictionary<string, string>(), string.Empty))
+
+                         // Filter out tags with empty keys, and trim whitespace
+                        .Where(kvp => !string.IsNullOrWhiteSpace(kvp.Key))
+                        .ToDictionary(kvp => kvp.Key.Trim(), kvp => kvp.Value?.Trim() ?? string.Empty);
+        }
+        else
+        {
+            globalTags = config
+                        .WithKeys(ConfigurationKeys.GlobalTags, "DD_TRACE_GLOBAL_TAGS")
+                        .AsDictionaryResult()
+                        .OverrideWith(
+                             RemapOtelTags(in otelTags),
+                             ErrorLog,
+                             () => new DefaultResult<IDictionary<string, string>>(new Dictionary<string, string>(), string.Empty))
+
+                         // Filter out tags with empty keys or empty values, and trim whitespace
+                        .Where(kvp => !string.IsNullOrWhiteSpace(kvp.Key) && !string.IsNullOrWhiteSpace(kvp.Value))
+                        .ToDictionary(kvp => kvp.Key.Trim(), kvp => kvp.Value.Trim());
+        }
+
+        Environment = config
+                     .WithKeys(ConfigurationKeys.Environment)
+                     .AsString();
+
+        // DD_ENV has precedence over DD_TAGS
+        Environment = GetExplicitSettingOrTag(Environment, globalTags, Tags.Env, ConfigurationKeys.Environment);
+
+        var otelServiceName = config.WithKeys(ConfigurationKeys.OpenTelemetry.ServiceName).AsStringResult();
+        var serviceName = config
+                         .WithKeys(ConfigurationKeys.ServiceName, "DD_SERVICE_NAME")
+                         .AsStringResult()
+                         .OverrideWith(in otelServiceName, ErrorLog);
+
+        // DD_SERVICE has precedence over DD_TAGS
+        serviceName = GetExplicitSettingOrTag(serviceName, globalTags, Tags.Service, ConfigurationKeys.ServiceName);
+
+        if (tracerSettings.IsRunningInCiVisibility)
+        {
+            // Set the service name if not set
+            var isUserProvidedTestServiceTag = true;
+            var ciVisServiceName = serviceName;
+            if (string.IsNullOrEmpty(serviceName))
+            {
+                // Extract repository name from the git url and use it as a default service name.
+                ciVisServiceName = TestOptimization.Instance.TracerManagement?.GetServiceNameFromRepository(CIEnvironmentValues.Instance.Repository);
+                isUserProvidedTestServiceTag = false;
+            }
+
+            globalTags[Ci.Tags.CommonTags.UserProvidedTestServiceTag] = isUserProvidedTestServiceTag ? "true" : "false";
+
+            // Normalize the service name
+            ciVisServiceName = NormalizerTraceProcessor.NormalizeService(ciVisServiceName);
+            if (ciVisServiceName != serviceName)
+            {
+                serviceName = ciVisServiceName;
+                telemetry.Record(ConfigurationKeys.ServiceName, serviceName, recordValue: true, ConfigurationOrigins.Calculated);
+            }
+        }
+
+        ServiceName = serviceName;
+
+        ServiceVersion = config
+                        .WithKeys(ConfigurationKeys.ServiceVersion)
+                        .AsString();
+
+        // DD_VERSION has precedence over DD_TAGS
+        ServiceVersion = GetExplicitSettingOrTag(ServiceVersion, globalTags, Tags.Version, ConfigurationKeys.ServiceVersion);
+
+        GitCommitSha = config
+                      .WithKeys(ConfigurationKeys.GitCommitSha)
+                      .AsString();
+
+        // DD_GIT_COMMIT_SHA has precedence over DD_TAGS
+        GitCommitSha = GetExplicitSettingOrTag(GitCommitSha, globalTags, Ci.Tags.CommonTags.GitCommit, ConfigurationKeys.GitCommitSha);
+
+        GitRepositoryUrl = config
+                          .WithKeys(ConfigurationKeys.GitRepositoryUrl)
+                          .AsString();
+
+        // DD_GIT_REPOSITORY_URL has precedence over DD_TAGS
+        GitRepositoryUrl = GetExplicitSettingOrTag(GitRepositoryUrl, globalTags, Ci.Tags.CommonTags.GitRepository, ConfigurationKeys.GitRepositoryUrl);
+
+        var otelTraceEnabled = config
+                              .WithKeys(ConfigurationKeys.OpenTelemetry.TracesExporter)
+                              .AsBoolResult(
+                                   value => string.Equals(value, "none", StringComparison.OrdinalIgnoreCase)
+                                                ? ParsingResult<bool>.Success(result: false)
+                                                : ParsingResult<bool>.Failure());
+        TraceEnabled = config
+                      .WithKeys(ConfigurationKeys.TraceEnabled)
+                      .AsBoolResult()
+                      .OverrideWith(in otelTraceEnabled, ErrorLog, defaultValue: true);
+
+        if (tracerSettings.AzureAppServiceMetadata?.IsUnsafeToTrace == true)
+        {
+            telemetry.Record(ConfigurationKeys.TraceEnabled, false, ConfigurationOrigins.Calculated);
+            TraceEnabled = false;
+        }
+
+        var disabledIntegrationNames = config.WithKeys(ConfigurationKeys.DisabledIntegrations)
+                                             .AsString()
+                                            ?.Split([';'], StringSplitOptions.RemoveEmptyEntries) ?? [];
+
+        // If Activity support is enabled, we shouldn't enable the OTel listener
+        DisabledIntegrationNames = tracerSettings.IsActivityListenerEnabled
+                                       ? new HashSet<string>(disabledIntegrationNames, StringComparer.OrdinalIgnoreCase)
+                                       : new HashSet<string>([..disabledIntegrationNames, nameof(IntegrationId.OpenTelemetry)], StringComparer.OrdinalIgnoreCase);
+
+        Integrations = new IntegrationSettingsCollection(source, DisabledIntegrationNames);
+        RecordDisabledIntegrationsTelemetry(Integrations, Telemetry);
+
+#pragma warning disable 618 // App analytics is deprecated, but still used
+        AnalyticsEnabled = config.WithKeys(ConfigurationKeys.GlobalAnalyticsEnabled)
+                                 .AsBool(defaultValue: false);
+#pragma warning restore 618
+
+#pragma warning disable 618 // this parameter has been replaced but may still be used
+        MaxTracesSubmittedPerSecond = config
+                                     .WithKeys(ConfigurationKeys.TraceRateLimit, ConfigurationKeys.MaxTracesSubmittedPerSecond)
+#pragma warning restore 618
+                                     .AsInt32(defaultValue: 100);
+
+        // mutate dictionary to remove without "env", "version", "git.commit.sha" or "git.repository.url" tags
+        // these value are used for "Environment" and "ServiceVersion", "GitCommitSha" and "GitRepositoryUrl" properties
+        // or overriden with DD_ENV, DD_VERSION, DD_GIT_COMMIT_SHA and DD_GIT_REPOSITORY_URL respectively
+        globalTags.Remove(Tags.Service);
+        globalTags.Remove(Tags.Env);
+        globalTags.Remove(Tags.Version);
+        globalTags.Remove(Ci.Tags.CommonTags.GitCommit);
+        globalTags.Remove(Ci.Tags.CommonTags.GitRepository);
+        GlobalTags = new(globalTags);
+
+        var headerTagsNormalizationFixEnabled = config
+                                               .WithKeys(ConfigurationKeys.FeatureFlags.HeaderTagsNormalizationFixEnabled)
+                                               .AsBool(defaultValue: true);
+
+        // Filter out tags with empty keys or empty values, and trim whitespaces
+        HeaderTags = InitializeHeaderTags(config, ConfigurationKeys.HeaderTags, headerTagsNormalizationFixEnabled) ?? ReadOnlyDictionary.Empty;
+
+        // Filter out tags with empty keys or empty values, and trim whitespaces
+        GrpcTags = InitializeHeaderTags(config, ConfigurationKeys.GrpcTags, headerTagsNormalizationFixEnabled: true) ?? ReadOnlyDictionary.Empty;
+
+        CustomSamplingRules = config.WithKeys(ConfigurationKeys.CustomSamplingRules).AsString();
+
+        GlobalSamplingRate = BuildSampleRate(ErrorLog, in config);
+
+        // We need to record a default value for configuration reporting
+        // However, we need to keep GlobalSamplingRateInternal null because it changes the behavior of the tracer in subtle ways
+        // (= we don't run the sampler at all if it's null, so it changes the tagging of the spans, and it's enforced by system tests)
+        if (GlobalSamplingRate is null)
+        {
+            Telemetry.Record(ConfigurationKeys.GlobalSamplingRate, 1.0, ConfigurationOrigins.Default);
+        }
+
+        StartupDiagnosticLogEnabled = config.WithKeys(ConfigurationKeys.StartupDiagnosticLogEnabled).AsBool(defaultValue: true);
+
+        KafkaCreateConsumerScopeEnabled = config
+                                         .WithKeys(ConfigurationKeys.KafkaCreateConsumerScopeEnabled)
+                                         .AsBool(defaultValue: true);
+        ServiceNameMappings = TracerSettings.InitializeServiceNameMappings(config, ConfigurationKeys.ServiceNameMappings) ?? ReadOnlyDictionary.Empty;
+
+        TracerMetricsEnabled = config
+                              .WithKeys(ConfigurationKeys.TracerMetricsEnabled)
+                              .AsBool(defaultValue: false);
+
+        var httpServerErrorStatusCodes = config
+#pragma warning disable 618 // This config key has been replaced but may still be used
+                                        .WithKeys(ConfigurationKeys.HttpServerErrorStatusCodes, ConfigurationKeys.DeprecatedHttpServerErrorStatusCodes)
+#pragma warning restore 618
+                                        .AsString(defaultValue: "500-599");
+
+        HttpServerErrorStatusCodes = ParseHttpCodesToArray(httpServerErrorStatusCodes);
+
+        var httpClientErrorStatusCodes = config
+#pragma warning disable 618 // This config key has been replaced but may still be used
+                                        .WithKeys(ConfigurationKeys.HttpClientErrorStatusCodes, ConfigurationKeys.DeprecatedHttpClientErrorStatusCodes)
+#pragma warning restore 618
+                                        .AsString(defaultValue: "400-499");
+
+        HttpClientErrorStatusCodes = ParseHttpCodesToArray(httpClientErrorStatusCodes);
+    }
+
+    // Settings that can be set via remote config
+
+    /// <summary>
+    /// Gets a value indicating whether tracing is enabled.
+    /// Default is <c>true</c>.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.TraceEnabled"/>
+    public bool TraceEnabled { get; }
+
+    /// <summary>
+    /// Gets a value indicating custom sampling rules.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.CustomSamplingRules"/>
+    public string? CustomSamplingRules { get; }
+
+    /// <summary>
+    /// Gets a value indicating a global rate for sampling.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.GlobalSamplingRate"/>
+    public double? GlobalSamplingRate { get; }
+
+    /// <summary>
+    /// Gets a value indicating whether correlation identifiers are
+    /// automatically injected into the logging context.
+    /// Default is <c>true</c>.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.LogsInjectionEnabled"/>
+    public bool LogsInjectionEnabled { get; }
+
+    /// <summary>
+    /// Gets the global tags, which are applied to all <see cref="Span"/>s.
+    /// </summary>
+    public ReadOnlyDictionary<string, string> GlobalTags { get; }
+
+    /// <summary>
+    /// Gets the map of header keys to tag names, which are applied to the root <see cref="Span"/>
+    /// of incoming and outgoing HTTP requests.
+    /// </summary>
+    public ReadOnlyDictionary<string, string> HeaderTags { get; }
+
+    // Additional settings that can be set in code
+
+    // TODO: This one will be hard, we should likely consider something completely different to handle these changes
+    // public const string AgentUriKey = "DD_TRACE_AGENT_URL";
+
+    // TODO: this one is a problem - it can be changed in code, but the data pipeline needs to be enabled/disabled based on it
+    // we will likely need to make a breaking change so that it _can't_ be changed in code
+    // public const string StatsComputationEnabledKey = "DD_TRACE_STATS_COMPUTATION_ENABLED";
+
+    /// <summary>
+    /// Gets a value indicating whether the diagnostic log at startup is enabled
+    /// </summary>
+    public bool StartupDiagnosticLogEnabled { get; }
+
+    /// <summary>
+    /// Gets the default environment name applied to all spans.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.Environment"/>
+    public string? Environment { get; }
+
+    /// <summary>
+    /// Gets the service name applied to top-level spans and used to build derived service names.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.ServiceName"/>
+    public string? ServiceName { get; }
+
+    /// <summary>
+    /// Gets the version tag applied to all spans.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.ServiceVersion"/>
+    public string? ServiceVersion { get; }
+
+    /// <summary>
+    /// Gets the names of disabled integrations.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.DisabledIntegrations"/>
+    public HashSet<string> DisabledIntegrationNames { get; }
+
+    /// <summary>
+    /// Gets the map of metadata keys to tag names, which are applied to the root <see cref="Span"/>
+    /// of incoming and outgoing GRPC requests.
+    /// </summary>
+    public ReadOnlyDictionary<string, string> GrpcTags { get; }
+
+    /// <summary>
+    /// Gets a value indicating whether internal metrics
+    /// are enabled and sent to DogStatsd.
+    /// </summary>
+    public bool TracerMetricsEnabled { get; }
+
+    /// <summary>
+    /// Gets a collection of <see cref="IntegrationSettings"/> keyed by integration name.
+    /// </summary>
+    public IntegrationSettingsCollection Integrations { get; }
+
+    /// <summary>
+    /// Gets a value indicating whether default Analytics are enabled.
+    /// Settings this value is a shortcut for setting
+    /// <see cref="Configuration.IntegrationSettings.AnalyticsEnabled"/> on some predetermined integrations.
+    /// See the documentation for more details.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.GlobalAnalyticsEnabled"/>
+    [Obsolete(DeprecationMessages.AppAnalytics)]
+    public bool AnalyticsEnabled { get; }
+
+    /// <summary>
+    /// Gets a value indicating the maximum number of traces set to AutoKeep (p1) per second.
+    /// Default is <c>100</c>.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.TraceRateLimit"/>
+    public int MaxTracesSubmittedPerSecond { get; }
+
+    /// <summary>
+    /// Gets a value indicating whether a span context should be created on exiting a successful Kafka
+    /// Consumer.Consume() call, and closed on entering Consumer.Consume().
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.KafkaCreateConsumerScopeEnabled"/>
+    public bool KafkaCreateConsumerScopeEnabled { get; }
+
+    /// <summary>
+    /// Gets the HTTP status code that should be marked as errors for server integrations.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.HttpServerErrorStatusCodes"/>
+    internal bool[] HttpServerErrorStatusCodes { get; }
+
+    /// <summary>
+    /// Gets the HTTP status code that should be marked as errors for client integrations.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.HttpClientErrorStatusCodes"/>
+    internal bool[] HttpClientErrorStatusCodes { get; }
+
+    /// <summary>
+    /// Gets configuration values for changing service names based on configuration
+    /// </summary>
+    internal ReadOnlyDictionary<string, string> ServiceNameMappings { get; }
+
+    // These ones can't be _directly_ changed, but are dependent on things that _can_
+    // so they are _implicitly_ dynamic
+
+    /// <summary>
+    /// Gets the application's git repository url.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.GitRepositoryUrl"/>
+    internal string? GitRepositoryUrl { get; }
+
+    /// <summary>
+    /// Gets the application's git commit hash.
+    /// </summary>
+    /// <seealso cref="ConfigurationKeys.GitCommitSha"/>
+    internal string? GitCommitSha { get; }
+
+    // Infra
+    internal OverrideErrorLog ErrorLog { get; }
+
+    internal IConfigurationTelemetry Telemetry { get; }
+
+    internal static ReadOnlyDictionary<string, string>? InitializeHeaderTags(ConfigurationBuilder config, string key, bool headerTagsNormalizationFixEnabled)
+    {
+        var configurationDictionary = config
+                                     .WithKeys(key)
+                                     .AsDictionary(allowOptionalMappings: true, defaultValue: null, "[]");
+
+        if (configurationDictionary == null)
+        {
+            return null;
+        }
+
+        var headerTags = new Dictionary<string, string>(configurationDictionary.Count);
+
+        foreach (var kvp in configurationDictionary)
+        {
+            var headerName = kvp.Key.Trim();
+
+            if (string.IsNullOrEmpty(headerName))
+            {
+                continue;
+            }
+
+            if (InitializeHeaderTag(tagName: kvp.Value, headerTagsNormalizationFixEnabled, out var finalTagName))
+            {
+                headerTags.Add(headerName, finalTagName);
+            }
+        }
+
+        return new(headerTags);
+    }
+
+    internal static bool InitializeHeaderTag(
+        string? tagName,
+        bool headerTagsNormalizationFixEnabled,
+        [NotNullWhen(true)] out string? finalTagName)
+    {
+        tagName = tagName?.Trim();
+
+        if (string.IsNullOrEmpty(tagName))
+        {
+            // The user did not provide a tag name. Normalization will happen later, when adding the tag prefix.
+            finalTagName = string.Empty;
+            return true;
+        }
+
+        if (!SpanTagHelper.IsValidTagName(tagName!, out tagName))
+        {
+            // invalid tag name
+            finalTagName = null;
+            return false;
+        }
+
+        if (headerTagsNormalizationFixEnabled)
+        {
+            // Default code path: if the user provided a tag name, don't try to normalize it.
+            finalTagName = tagName;
+            return true;
+        }
+
+        // user opted via feature flag into the previous behavior,
+        // where tag names were normalized even when specified
+        // (but _not_ spaces, due to a bug in the normalization code)
+        return SpanTagHelper.TryNormalizeTagName(tagName, normalizeSpaces: false, out finalTagName);
+    }
+
+    internal static bool[] ParseHttpCodesToArray(string httpStatusErrorCodes)
+    {
+        bool[] httpErrorCodesArray = new bool[600];
+
+        void TrySetValue(int index)
+        {
+            if (index >= 0 && index < httpErrorCodesArray.Length)
+            {
+                httpErrorCodesArray[index] = true;
+            }
+        }
+
+        string[] configurationsArray = httpStatusErrorCodes.Replace(" ", string.Empty).Split(',');
+
+        foreach (string statusConfiguration in configurationsArray)
+        {
+            int startStatus;
+
+            // Checks that the value about to be used follows the `401-404` structure or single 3 digit number i.e. `401` else log the warning
+            if (!Regex.IsMatch(statusConfiguration, @"^\d{3}-\d{3}$|^\d{3}$"))
+            {
+                // TODO: this should be logged in telemetry or the ErrorLog override or something
+                // Log.Warning("Wrong format '{0}' for DD_TRACE_HTTP_SERVER/CLIENT_ERROR_STATUSES configuration.", statusConfiguration);
+            }
+
+            // If statusConfiguration equals a single value i.e. `401` parse the value and save to the array
+            else if (int.TryParse(statusConfiguration, out startStatus))
+            {
+                TrySetValue(startStatus);
+            }
+            else
+            {
+                string[] statusCodeLimitsRange = statusConfiguration.Split('-');
+
+                startStatus = int.Parse(statusCodeLimitsRange[0]);
+                int endStatus = int.Parse(statusCodeLimitsRange[1]);
+
+                if (endStatus < startStatus)
+                {
+                    startStatus = endStatus;
+                    endStatus = int.Parse(statusCodeLimitsRange[0]);
+                }
+
+                for (int statusCode = startStatus; statusCode <= endStatus; statusCode++)
+                {
+                    TrySetValue(statusCode);
+                }
+            }
+        }
+
+        return httpErrorCodesArray;
+    }
+
+    internal bool IsErrorStatusCode(int statusCode, bool serverStatusCode)
+    {
+        var source = serverStatusCode ? HttpServerErrorStatusCodes : HttpClientErrorStatusCodes;
+
+        if (source == null)
+        {
+            return false;
+        }
+
+        if (statusCode >= source.Length)
+        {
+            return false;
+        }
+
+        return source[statusCode];
+    }
+
+    internal bool IsIntegrationEnabled(IntegrationId integration, bool defaultValue = true)
+    {
+        if (TraceEnabled && !_domainMetadata.ShouldAvoidAppDomain())
+        {
+            return Integrations[integration].Enabled ?? defaultValue;
+        }
+
+        return false;
+    }
+
+    [Obsolete(DeprecationMessages.AppAnalytics)]
+    internal double? GetIntegrationAnalyticsSampleRate(IntegrationId integration, bool enabledWithGlobalSetting)
+    {
+        var integrationSettings = Integrations[integration];
+        var analyticsEnabled = integrationSettings.AnalyticsEnabled ?? (enabledWithGlobalSetting && AnalyticsEnabled);
+        return analyticsEnabled ? integrationSettings.AnalyticsSampleRate : (double?)null;
+    }
+
+    private static ConfigurationBuilder.ClassConfigurationResultWithKey<IDictionary<string, string>> RemapOtelTags(
+        in ConfigurationBuilder.ClassConfigurationResultWithKey<IDictionary<string, string>> original)
+    {
+        if (original.ConfigurationResult is { IsValid: true, Result: { } values })
+        {
+            // Update well-known service information resources
+            if (values.TryGetValue("deployment.environment", out var envValue))
+            {
+                values.Remove("deployment.environment");
+                values[Tags.Env] = envValue;
+            }
+
+            if (values.TryGetValue("service.name", out var serviceValue))
+            {
+                values.Remove("service.name");
+                values[Tags.Service] = serviceValue;
+            }
+
+            if (values.TryGetValue("service.version", out var versionValue))
+            {
+                values.Remove("service.version");
+                values[Tags.Version] = versionValue;
+            }
+        }
+
+        return original;
+    }
+
+    private static void RecordDisabledIntegrationsTelemetry(IntegrationSettingsCollection integrations, IConfigurationTelemetry telemetry)
+    {
+        // Record the final disabled settings values in the telemetry, we can't quite get this information
+        // through the IntegrationTelemetryCollector currently so record it here instead
+        StringBuilder? sb = null;
+
+        foreach (var setting in integrations.Settings)
+        {
+            if (setting.Enabled == false)
+            {
+                sb ??= StringBuilderCache.Acquire();
+                sb.Append(setting.IntegrationName);
+                sb.Append(';');
+            }
+        }
+
+        var value = sb is null ? null : StringBuilderCache.GetStringAndRelease(sb);
+        telemetry.Record(ConfigurationKeys.DisabledIntegrations, value, recordValue: true, ConfigurationOrigins.Calculated);
+    }
+
+    private static double? BuildSampleRate(OverrideErrorLog log, in ConfigurationBuilder config)
+    {
+        // The "overriding" is complex, so we can't use the usual `OverrideWith()` approach
+        var ddSampleRate = config.WithKeys(ConfigurationKeys.GlobalSamplingRate).AsDoubleResult();
+        var otelSampleType = config.WithKeys(ConfigurationKeys.OpenTelemetry.TracesSampler).AsStringResult();
+        var otelSampleRate = config.WithKeys(ConfigurationKeys.OpenTelemetry.TracesSamplerArg).AsDoubleResult();
+
+        double? ddResult = ddSampleRate.ConfigurationResult.IsValid ? ddSampleRate.ConfigurationResult.Result : null;
+
+        // more complex, so can't use built-in `Merge()` support
+        if (ddSampleRate.ConfigurationResult.IsPresent)
+        {
+            if (otelSampleType.ConfigurationResult.IsPresent)
+            {
+                log.LogDuplicateConfiguration(ddSampleRate.Key, otelSampleType.Key);
+            }
+
+            if (otelSampleRate.ConfigurationResult.IsPresent)
+            {
+                log.LogDuplicateConfiguration(ddSampleRate.Key, otelSampleRate.Key);
+            }
+        }
+        else if (otelSampleType.ConfigurationResult is { IsValid: true, Result: { } samplerName })
+        {
+            const string parentbasedAlwaysOn = "parentbased_always_on";
+            const string parentbasedAlwaysOff = "parentbased_always_off";
+            const string parentbasedTraceidratio = "parentbased_traceidratio";
+
+            string? supportedSamplerName = samplerName switch
+            {
+                parentbasedAlwaysOn => parentbasedAlwaysOn,
+                "always_on" => parentbasedAlwaysOn,
+                parentbasedAlwaysOff => parentbasedAlwaysOff,
+                "always_off" => parentbasedAlwaysOff,
+                parentbasedTraceidratio => parentbasedTraceidratio,
+                "traceidratio" => parentbasedTraceidratio,
+                _ => null,
+            };
+
+            if (supportedSamplerName is null)
+            {
+                log.EnqueueAction(
+                    (log, _) =>
+                    {
+                        log.Warning(
+                            "OpenTelemetry configuration {OpenTelemetryConfiguration}={OpenTelemetryValue} is not supported. Using default configuration.",
+                            otelSampleType.Key,
+                            samplerName);
+                    });
+                return ddResult;
+            }
+
+            if (!string.Equals(samplerName, supportedSamplerName, StringComparison.OrdinalIgnoreCase))
+            {
+                log.LogUnsupportedConfiguration(otelSampleType.Key, samplerName, supportedSamplerName);
+            }
+
+            var openTelemetrySampleRateResult = supportedSamplerName switch
+            {
+                parentbasedAlwaysOn => ConfigurationResult<double>.Valid(1.0),
+                parentbasedAlwaysOff => ConfigurationResult<double>.Valid(0.0),
+                parentbasedTraceidratio => otelSampleRate.ConfigurationResult,
+                _ => ConfigurationResult<double>.ParseFailure(),
+            };
+
+            if (openTelemetrySampleRateResult is { Result: { } sampleRateResult, IsValid: true })
+            {
+                return sampleRateResult;
+            }
+
+            log.LogInvalidConfiguration(otelSampleRate.Key);
+        }
+
+        return ddResult;
+    }
+
+    private string? GetExplicitSettingOrTag(
+        string? explicitSetting,
+        Dictionary<string, string> globalTags,
+        string tag,
+        string telemetryKey)
+    {
+        string? result = null;
+        if (!string.IsNullOrWhiteSpace(explicitSetting))
+        {
+            result = explicitSetting!.Trim();
+            if (result != explicitSetting)
+            {
+                Telemetry.Record(telemetryKey, result, recordValue: true, ConfigurationOrigins.Calculated);
+            }
+        }
+        else
+        {
+            var version = globalTags.GetValueOrDefault(tag);
+            if (!string.IsNullOrWhiteSpace(version))
+            {
+                result = version.Trim();
+                Telemetry.Record(telemetryKey, result, recordValue: true, ConfigurationOrigins.Calculated);
+            }
+        }
+
+        return result;
+    }
+}
